feat: Add deletePost method to PostController
This commit adds the `deletePost` method to the `PostController` class in `post.controller.ts`. The `deletePost` method is decorated with `@Delete(':id')` and `@CreateRpcPayload()` to handle deleting a post. It calls the `deletePost` method of the `PostService` to delete the post. This change enables the functionality to delete a post in the application.